- include the correct thrift include dir when building scribe. still got an error.

+++ hooks/fb303.py
@@ -2,33 +2,42 @@ import logging
 import os
 import fileinput
 
-log = logging.getLogger('fb303-hook')
+log = logging.getLogger('Makefile-hook')
 
-def premake(options, buildout):
-    # import pdb; pdb.set_trace()
+def premake_303(options, buildout):
+    return _premake(options, buildout, 'prefix', '/usr/local', 
+            buildout['thrift']['location'])
+
+def premake_scribe(options, buildout):
+    return _premake(options, buildout, 'fb303_home', '/usr/local', 
+            buildout['thrift']['location'])
+
+def _premake(options, buildout, what, prefix_pattern, replacement):
     basedir = options['path']
     makefile_list = [os.path.join(basedir, "Makefile"), 
             os.path.join(basedir, "cpp", "Makefile"),
-            os.path.join(basedir, "py", "Makefile")
+            os.path.join(basedir, "py", "Makefile"),
+            os.path.join(basedir, "src", "Makefile"),
             ]
 
     log.debug('pre-make-hook on %s' % basedir)
 
     for makefile in makefile_list:
-        log.info("fixing path in ``%s``" % makefile)
-        
         try:
             os.remove(makefile+"~")
         except:
             pass;
-        os.rename(makefile, makefile+"~")
-        dst = open(makefile, "w")
-        src = open(makefile+"~", "r")
-        
-        for line in src:
-            if "prefix = /usr/local" in line:
-                log.debug("... %s" % line)
-                line=line.replace('/usr/local', buildout['thrift']['location'])
-                log.info(">>> %s" % line)
-            dst.write(line)
-        fileinput.close()
+
+        if os.path.exists(makefile):
+            log.info("fixing path in ``%s``" % makefile)
+            os.rename(makefile, makefile+"~")
+            dst = open(makefile, "w")
+            src = open(makefile+"~", "r")
+            
+            for line in src:
+                if "%s = %s" % (what, prefix_pattern) in line:
+                    log.debug("... %s" % line)
+                    line=line.replace(prefix_pattern, replacement)
+                    log.info(">>> %s" % line)
+                dst.write(line)
+            fileinput.close()
